Update card so that image appears below heading
This is going to need some further testing but I've checked it
with headingsMap and with VoiceOver and it seems the combination of
using a screenreader only class to hide the visual heading and then
using aria-hidden to hide the visual heading from AT isn't introducing
problems. This might need to be revisted if further testing reveals
issues.